move state and effect to inputRepo component, make ScrollableDropDown generic as possible
