+++ src/components/ScrollableDropDown.jsx
@@ -1,31 +1,13 @@
-import { useEffect, useState } from "react";
 import AsyncSelect from "react-select/async";
 
-export default function ScrollableDropDown({ repositories, loadingStatus }) {
-  const [reposName, setReposName] = useState([]);
+export default function ScrollableDropDown({ options, loadingStatus }) {
   const filterOptions = (inputValue) => {
-    const repos = reposName;
-    const selectedOption = repos.filter((i) =>
+    const mOptions = options;
+    const selectedOption = mOptions.filter((i) =>
       i.label.toLowerCase().includes(inputValue.toLowerCase())
     );
     return selectedOption;
   };
-
-  useEffect(() => {
-    let repositoriesArray = [];
-    if (repositories && repositories.length > 0) {
-      repositoriesArray = repositories.map((repository) => {
-        return { value: repository.name, label: repository.name };
-      });
-      console.log(
-        "ReposName: " +
-          JSON.stringify(repositoriesArray) +
-          " " +
-          repositoriesArray.length
-      );
-      setReposName(repositoriesArray);
-    }
-  }, [repositories]);
 
   const loadOptions = (inputValue, callback) => {
     setTimeout(() => {
@@ -37,8 +19,8 @@ export default function ScrollableDropDown({ repositories, loadingStatus }) {
     <AsyncSelect
       cacheOptions
       loadOptions={loadOptions}
-      defaultOptions={reposName}
-      defaultValue={reposName[0]}
+      defaultOptions={options}
+      defaultValue={options[0]}
       loadingStatus={loadingStatus}
     />
   );
